added confirmation popup in all components delete buttons

+++ src/Components/Pages/Dashboard/Reviews/Reviews.js
@@ -28,20 +28,25 @@ const Reviews = () => {
     }, [reviews])
 
     const handleRemoveOrder = (id) => {
-        fetch(`https://fragrance-shop.herokuapp.com/reviews/${id}`, {
-            method: 'DELETE',
-            headers: {
-                'authorization': `Bearer ${token}`
-            }
-        })
-            .then(res => res.json())
-            .then(data => {
-                if (data.acknowledged) {
-                    console.log(data);
-                    setSuccessdeleting(true);
-                    setLoading(false);
+        const proceed = window.confirm(`Are you sure, to Delete this order ${id} ?`);
+        if (proceed) {
+            fetch(`https://fragrance-shop.herokuapp.com/reviews/${id}`, {
+                method: 'DELETE',
+                headers: {
+                    'authorization': `Bearer ${token}`
                 }
             })
+                .then(res => res.json())
+                .then(data => {
+                    if (data.acknowledged) {
+                        console.log(data);
+                        setSuccessdeleting(true);
+                        setLoading(false);
+                    }
+                })
+        } else {
+            setLoading(false)
+        }
     }
     return (
         <div>
